Add user sign-out functionality
Implemented the ability for users to sign out of their accounts. Added the 'logOut' action in the actions module and created a new 'SignOutBTN' component for the account page. Adjusted the account page layout to incorporate the new sign-out button. Updated authentication library exports to include 'signOut' method to support this feature.